+++ spec/spec_helper.rb
@@ -1,6 +1,5 @@
-require 'rspec/expectations'
 require 'chefspec'
 require 'chefspec/berkshelf'
-require 'chefspec/server'
+require_relative '../libraries/default'
 
 at_exit { ChefSpec::Coverage.report! }